requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
timm
Pillow
python-multipart
starlette
pydantic
pydantic-settings
ultralytics
onnxruntime
numpy
sqlalchemy
pymysql
cryptography
sqlalchemy_utils
python-dotenv
google-genai
protobuf
alembic